rearrange: move validations from utils to validations
